# [Database] Compress values stored in badgerDB