--- a/storage/badger/operation/modifiers_test.go
+++ b/storage/badger/operation/modifiers_test.go
@@ -10,7 +10,6 @@ import (
 	"github.com/dgraph-io/badger/v2"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/vmihailenco/msgpack/v4"
 
 	"github.com/onflow/flow-go/storage"
 	"github.com/onflow/flow-go/utils/unittest"
@@ -20,7 +19,7 @@ func TestSkipDuplicates(t *testing.T) {
 	unittest.RunWithBadgerDB(t, func(db *badger.DB) {
 		e := Entity{ID: 1337}
 		key := []byte{0x01, 0x02, 0x03}
-		val, _ := msgpack.Marshal(e)
+		val, _ := encodeAndCompress(e)
 
 		// persist first time
 		err := db.Update(insert(key, e))