Close https://github.com/onflow/flow-go/issues/5402

## Summary
This PR uses snappy to compress the data stored in badgerDB. Benchmark shows a significant save on storage especially for chunk data pack storage, which takes more than 90% of disk usage.

## First attempt
Initially I [tried the Snappy options from badger](https://github.com/onflow/flow-go/pull/4495), however, the benchmark shows no difference, the data is still uncompressed. 

## Second attempt
In order to compress the data, I need to manually compress the value stored in badger. I chose the Snappy algorithm, as it doesn't sacrifices the speed too much and still have a good compression rate.

## Benchmark Storage
I ran the localnet instance with benchmark tool sending transactions for 5 mins. The result shows the protocol data is reduced 10%, and chunk data pack is reduced 65%:
```
# Without compression
execution protocol data: 4.4MB
execution chunk data pack: 124MB

# With compression
execution protocol data: 4MB
execution chunk data pack data: 43MB
```

## Benchmark Speed of pure encoding and decoding
Pure encoding is 28% slower, Pure decoding is 3.6 times slower:
```
# encoding without compression
BenchmarkEncodeWithoutCompress-10         726984              1642 ns/op
# encoding with compression
BenchmarkEncodeAndCompress-10             531303              2272 ns/op

# decoding without compression
BenchmarkDecodeWithoutUncompress-10      6708643               172.6 ns/op
# decoding with compression
BenchmarkDecodeUncompressed-10           1922100               624.4 ns/op
```

## Benchmark Speed of round trip saving and reading from database
It shows 3% slowness in read, and 5% slowness in write:
```
# encoding without compression and saving to database
BenchmarkReadResult-10            478755              2316 ns/op
# encoding with compression and saving to database
BenchmarkReadResult-10            559629              2437 ns/op

# reading from database and decoding without compression
BenchmarkSaveResult-10             25435             45273 ns/op
# reading from database and decoding with compression
BenchmarkSaveResult-10             24032             46323 ns/op
```