@fxamacker 2024-03-04
Maybe add memory usage to the benchmarks posted here: `B/op` and `allocs/op` columns.