@bluesign 2024-03-04
> Initially I https://github.com/onflow/flow-go/pull/4495, however, the benchmark shows no difference, the data is still uncompressed.

Probably it was still in vlog ( not compacted ) badger is compressing when compacting. (in background) But this approach is much better, other one is really hard to benchmark.