@zhangchiqing 2024-04-10
Cose for now, since I ran into an issue implementing the migration of all the badger values to compressed format. Will reopen when https://github.com/onflow/flow-go/pull/5627 is open and merged.